+++ views/answer-add.pug
@@ -5,10 +5,10 @@ include utils.pug
 block content
   h1 #{formTitle}
   +validationErrorSummary(errors)
-  form(action='/answer/add' method='POST')
+  form(action=`/questions/${question.id}/answer/add` method='POST')
     input(type='hidden' name='_csrf' value=csrfToken)
     label(for='content') Content:
     textarea(name='content' value=content)
     div.form-buttons
       button(type='submit' class='form-submit-button') Add Answer
-      a(href='/questions/#{question.id}' class='form-cancel-button') Cancel
+      a(href=`/questions/${question.id}` class='form-cancel-button') Cancel
