+++ views/questions.pug
@@ -25,3 +25,4 @@ block content
                             span.username-span= `${question.User.username}`
                             .modified-div
                             span.modified-span= `created at ${question.createdAt}`
+    script(src="/javascripts/votes.js")
